change primary constructor PostsAdapter

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -6,9 +6,9 @@ data class Post(
     val content: String = "", //Текст сообщения
     val published: String = "", //Дата и время публикации
     val likedByMe: Boolean = false,  //Лайк включен или отключен
-    val likeClickCount: Int = 0, //Счётчик лайков
-    val shareClickCount: Int = 0, //Счётчик репостов
-    val lookClickCount: Int = 0 //Счётчик просмотров
+    val likeClickCount: Long = 0L, //Счётчик лайков
+    val shareClickCount: Long = 0L, //Счётчик репостов
+    val lookClickCount: Long = 0L, //Счётчик просмотров
 )
 //Значения по умолчанию добавлены, чтобы было проще писать тесты
 // (не указывать в каждом тесте все значения класса)
